LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-attribute-basic.tentative-expected.txt

Pop upPop upPop upPop upDifferent element typeDifferent element typeDifferent element typeInvalid popover value - defaults to popover=manualInvalid popover value - defaults to popover=manualInvalid popover value - defaults to popover=manual
Not a popover
Dialog without popover attribute
Animated popover

FAIL The element <div popover="" id="boolean">Pop up</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div popover="">Pop up</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div popover="auto">Pop up</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div popover="manual">Pop up</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <article popover="">Different element type</article> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <header popover="">Different element type</header> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <nav popover="">Different element type</nav> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <input type="text" popover="" value="Different element type"> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <dialog popover="">Dialog with popover attribute</dialog> should behave as a popover. The string did not match the expected pattern.
FAIL The element <dialog popover="manual">Dialog with popover=manual</dialog> should behave as a popover. The string did not match the expected pattern.
FAIL The element <div popover="true">Invalid popover value - defaults to popover=manual</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div popover="popover">Invalid popover value - defaults to popover=manual</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div popover="invalid">Invalid popover value - defaults to popover=manual</div> should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL The element <div>Not a popover</div> should *not* behave as a popover. The string did not match the expected pattern.
FAIL The element <dialog open="">Dialog without popover attribute</dialog> should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <a popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <a> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <abbr popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <abbr> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <address popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <address> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <area popover> element should behave as a popover. The string did not match the expected pattern.
FAIL A <area> element should *not* behave as a popover. assert_equals: A non-popover should start out visible: Expected this element to be visible expected true but got false
FAIL A <article popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <article> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <aside popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <aside> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <b popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <b> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <bdi popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <bdi> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <bdo popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <bdo> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <blockquote popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <blockquote> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <body popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <body> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <button popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <button> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <canvas popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <canvas> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <caption popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <caption> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <cite popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <cite> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <code popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <code> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <col popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <col> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <colgroup popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <colgroup> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <data popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <data> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <dd popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <dd> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <del popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <del> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <details popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <details> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <dfn popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <dfn> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <div popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <div> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <dl popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <dl> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <dt popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <dt> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <em popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <em> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <fieldset popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <fieldset> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <figcaption popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <figcaption> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <figure popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <figure> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <footer popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <footer> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <form popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <form> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h1 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h1> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h2 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h2> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h3 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h3> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h4 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h4> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h5 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h5> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <h6 popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <h6> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <header popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <header> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <hr popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <hr> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <html popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <html> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <i popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <i> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <iframe popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <iframe> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <img popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <img> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <input popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <input> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <ins popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <ins> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <kbd popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <kbd> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <label popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <label> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <legend popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <legend> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <li popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <li> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <main popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <main> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <map popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <map> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <mark popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <mark> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <menu popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <menu> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <meter popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <meter> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <nav popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <nav> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <object popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <object> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <ol popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <ol> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <optgroup popover> element should behave as a popover. assert_equals: A popover should start out hidden: Non-showing popovers should have display:none expected "none" but got "block"
FAIL A <optgroup> element should *not* behave as a popover. assert_equals: A non-popover should start out visible: Expected this element to be visible expected true but got false
FAIL A <option popover> element should behave as a popover. assert_equals: A popover should start out hidden: Non-showing popovers should have display:none expected "none" but got "block"
FAIL A <option> element should *not* behave as a popover. assert_equals: A non-popover should start out visible: Expected this element to be visible expected true but got false
FAIL A <output popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <output> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <p popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <p> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <pre popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <pre> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <progress popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <progress> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <q popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <q> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <rt popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <rt> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <ruby popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <ruby> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <s popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <s> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <samp popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <samp> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <section popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <section> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <select popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <select> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <small popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <small> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <source popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <source> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <span popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <span> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <strong popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <strong> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <sub popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <sub> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <sup popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <sup> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <summary popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <summary> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <table popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <table> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <tbody popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <tbody> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <td popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <td> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <textarea popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <textarea> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <tfoot popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <tfoot> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <th popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <th> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <thead popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <thead> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <time popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <time> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <tr popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <tr> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <track popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <track> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <u popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <u> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <ul popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <ul> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <var popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <var> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL A <video popover> element should behave as a popover. assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL A <video> element should *not* behave as a popover. The string did not match the expected pattern.
FAIL IDL attribute reflection assert_equals: expected (string) "auto" but got (undefined) undefined
FAIL Popover attribute value should be case insensitive assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL Changing attribute values for popover should work assert_equals: A popover should start out hidden: Expected this element to be not visible expected false but got true
FAIL Changing attribute values should close open popovers popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL Removing a visible popover=auto element from the document should close the popover popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL A showing popover=auto does not match :modal popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL Removing a visible popover=manual element from the document should close the popover popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL A showing popover=manual does not match :modal popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL Changing the popover type in a "beforetoggle" event handler should throw an exception (during showPopover()) other_popover.showPopover is not a function. (In 'other_popover.showPopover()', 'other_popover.showPopover' is undefined)
FAIL Changing the popover type in a "beforetoggle" event handler should throw an exception (during hidePopover()) popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)
FAIL Changing a popover from auto to auto (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via attr), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via attr), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via attr), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via attr), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via attr), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to auto (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to manual (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to invalid (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to null (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from auto to undefined (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to auto (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to manual (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to invalid (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to null (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via idl), and then auto during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via idl), and then manual during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via idl), and then invalid during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via idl), and then null during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Changing a popover from manual to undefined (via idl), and then undefined during 'beforetoggle' works promise_test: Unhandled rejection with value: object "TypeError: popover.showPopover is not a function. (In 'popover.showPopover()', 'popover.showPopover' is undefined)"
FAIL Exceptions are thrown even when show/hide are animated assert_false: expected false got true
